[shell/linux] implement workqueue

+++ source/shell/sys/linux/include/linux/err.h
@@ -1,17 +1,17 @@
-#include <err.h>
+#include <linux/errno.h>
 
 #define MAX_ERRNO               4095
 
-#define IS_ERR_VALUE(__X)       ((unsigned long)(__X) >= (unsigned long)-MAX_ERRNO)
+#define IS_ERR_VALUE(__X)       ((unsigned long)(uintptr_t)(__X) >= (unsigned long)-MAX_ERRNO)
 
 static inline void * ERR_PTR(long error)
 {
-    return (void *)error;
+    return (void *)(uintptr_t)error;
 }
 
 static inline long PTR_ERR(const void *ptr)
 {
-    return (long) ptr;
+    return (long)(uintptr_t)ptr;
 }
 
 static inline bool IS_ERR(const void *ptr)
